segmentação do código para otimização
