+++ dictionary.py
@@ -1,0 +1,22 @@
+# pesquisa comum
+items = ["Tibia Coins", "Gold Token", "Silver Token","Alicorn Ring", "Amulet of Theurgy", "Arboreal Crown", "Arboreal Ring", "Arboreal Tome", "Arcanomancer Folio", "Arcanomancer Regalia",
+"Arcanomancer Sigil","Bloody Pincers", "Blueberry Cupcake", "Bone Fiddle", "Brimstone Fangs", "Broken Key Ring", "Broken Shamanic Staff", "Charged Alicorn Ring", "Charged Arboreal Ring",
+"Charged Arcanomancer Sigil", "Charged Spiritthorn Ring", "Cobra Axe", "Cobra Boots", "Cobra Club", "Cobra Crossbow", "Cobra Hood", "Cobra Rod", "Cobra Sword", "Cobra Wand", 
+"Cultish Mask", "Dawnfire Pantaloons", "Dawnfire Sherwani", "Decorative Ribbon","Dream Shroud", "Dwarven Armor", "Eldritch Bow", "Eldritch Breeches", "Eldritch Claymore", "Eldritch Cowl",
+"Eldritch Cuirass", "Eldritch Folio", "Eldritch Greataxe", "Eldritch Hood", "Eldritch Quiver", "Eldritch Rod", "Eldritch Shield", "Eldritch Tome", "Eldritch Wand", "Eldritch Warmace", "Elven Hoof", 
+"Elven Mail", "Elvish Talisman", "Enchanted Pendulet", "Enchanted Sleep Shawl", "Enchanted Theurgic Amulet", "Enchanted Turtle Amulet", "Exalted Core", "Exotic Amulet", "Exotic Legs", "Fabulous Legs",
+"Falcon Battleaxe", "Falcon Bow", "Falcon Circlet", "Falcon Coif", "Falcon Escutcheon", "Falcon Greaves", "Falcon Longsword","Falcon Mace", "Falcon Plate", "Falcon Rod", "Falcon Shield", 
+"Falcon Wand", "Feverbloom Boots", "Final Judgement", "Four-leaf Clover", "Foxtail", "Frostflower Boots", "Galea Mortis", "Ghost Chestplate", "Giant Shrimp", "Gloom Wolf Fur", "Gnome Armor", 
+"Gnome Helmet", "Gnome Legs", "Golden Can of Oil", "Gooey Mass", "Goosebump Leather", "Green Dragon Leather", "Library Ticket", "Lion Amulet", "Lion Axe", "Lion Hammer", "Lion Longbow", 
+"Lion Plate", "Lion Rod", "Lion Shield", "Lion Spangenhelm", "Lion Spellbook", "Lion Wand", "Little bowl of Myrrh", "Make-do Boots", "Midnight Sarong", "Midnight Tunic", 
+"Minor Crystalline Token", "Mooh'tah Shell", "Moohtant Horn", "Mutant Bone Boots", "Mutant Bone Kilt", "Mutated Skin Armor", "Mutated Skin Legs", "Naga Axe", "Naga Club", "Naga Crossbow", 
+"Naga Quiver", "Naga Rod", "Naga Wand", "Nightmare Horn", "Ornate Chestplate", "Pair of Dreamwalkers", "Pair of Soulstalkers", "Pair of Soulwalkers", "Perfect Behemoth Fang", 
+"Piece of Dead Brain", "Piece of Swampling Wood", "Prismatic Ring", "Protective Charm", "Rope Belt", "Rotworm Stew", "Sabretooth", "Sanguine Boots", "Sanguine Galoshes", "Sanguine Greaves", 
+"Sanguine Legs", "Scrubbing Brush","Shadow Cowl", "Shoulder Plate", "Silencer Claws",  "Sliver", "Slug Drug", "Snake Skin", "Soap", "Some Grimeleech Wings", "Soulbastion", 
+"Soulbleeder", "Soulcrusher", "Soulcutter", "Souleater", "Soulhexer", "Soulmaimer", "Soulmantle", "Soulpiercer", "Soulshanks", "Soulshell", "Soulshredder", "Soulshroud", "Soulstrider", 
+"Soultainter", "Spirit Guide", "Spiritthorn Armor", "Spiritthorn Helmet", "Stitched Mutant Hide Legs", "Strand of Medusa Hair", "Strawberry Cupcake", "Suspicious Device", 
+"Sweet Mangonaise Elixir", "The Cobra Amulet", "Thick Fur", "Toga Mortis", "Turtle Shell", "Umbral Master Spellbook", "Vampire Teeth", "Winter Wolf Fur", "Wyvern Talisman"]
+
+# pesquisa o segundo item (item grand sanguine e enchanted aparece primeiro)
+items_grand_sanguine = ["Pendulet", "Sleep Shawl", "Turtle Amulet", "Spiritthorn Ring", "Sanguine Battleaxe", "Sanguine Blade",  "Sanguine Bludgeon", "Sanguine Bow", "Sanguine Coil", "Sanguine Crossbow",
+"Sanguine Cudgel", "Sanguine Hatchet", "Sanguine Razor", "Sanguine Rod"]
